% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district_options.R
\name{district_options}
\alias{district_options}
\title{List all options for \code{district}}
\usage{
district_options()
}
\description{
Return a character vector for all valid
options for the \code{district} argument in \code{generate_ball()}
}
