% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_options.R
\name{expression_options}
\alias{expression_options}
\title{List all options for \code{expression}}
\usage{
expression_options()
}
\description{
Return a character vector for all valid
options for the \code{expression} argument in \code{generate_ball()}
}
