% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ball.R
\name{generate_ball}
\alias{generate_ball}
\title{Generate a basic District Ball image}
\usage{
generate_ball(
  district = "blank",
  expression = "happy",
  scale = 200,
  vjust = 1,
  hjust = 1,
  bg_colour = "#FFFFFF",
  caption = "",
  caption_colour = "#000000",
  caption_size = 20
)
}
\arguments{
\item{district}{character string specifying the district. Run \code{district_options()} to
find out all the valid options for this argument.}

\item{expression}{character string specifying the facial expression.
Run \code{expression_options()} to find out all the valid options for this argument.}

\item{scale}{numeric value. Size of the image to scale to. Defaults to 200.}

\item{vjust}{numeric value. Adjust the vertical position of the expression.}

\item{hjust}{numeric value. Adjust the horizontal of the expression.}

\item{bg_colour}{character string. HEX code for the background colour.}

\item{caption}{character string. Caption to appear at top of image. Defaults to blank.}

\item{caption_colour}{character string. HEX code for the caption colour.}

\item{caption_size}{numeric value. Text size of the caption.}
}
\description{
Generate a basic District Ball image by
specifying parameters for the district and the expression of the ball.
}
\examples{
\dontrun{
# Run a Kowloon City ball, fierce expression, scale to 500
generate_ball("kowloon-city",
              "fierce",
              scale = 500,
              vjust = 1.0,
              caption = "We need cleaner streets!",
              caption_size = 30)
}

}
