% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ball.R
\name{generate_ball}
\alias{generate_ball}
\title{Generate a basic District Ball image}
\usage{
generate_ball(
  district = "blank",
  expression = "happy",
  scale = 200,
  vjust = 1,
  caption = "",
  caption_size = 20
)
}
\arguments{
\item{district}{character string specifying the district. Valid inputs include "central-and-western",
"eastern", "islands", "kowloon-city", "kwai-tsing", "kwun-tong", "north", "sai-kung", "sha-tin", "sham-shui-po",
"southern", "tai-po", "tsuen-wan", "tuen-mun", "wan-chai", "wong-tai-sin", "yau-tsim-mong", and "yuen-long".
Defaults to "blank".}

\item{expression}{character string specifying the facial expression. Valid inputs include "fierce",
"happy", "normal", "sad", "skeptical".}

\item{scale}{numeric value. Size of the image to scale to. Defaults to 200.}

\item{vjust}{numeric value. Adjust the height of the expression.}

\item{caption}{character string. Caption to appear at top of image. Defaults to blank.}

\item{caption_size}{numeric value. Text size of the caption.}
}
\description{
Generate a basic District Ball image by
specifying parameters for the district and the expression of the ball.
}
\examples{
\dontrun{
# Run a Kowloon City ball, fierce expression, scale to 500
generate_ball("kowloon-city",
              "fierce",
              scale = 500,
              vjust = 1.0,
              caption = "We need cleaner streets!",
              caption_size = 30)
}

}
